Workaround for re-processing NFC intent
When onResume is called a second time after a tag was scanned (when the app was on background), the original NDEF_DISCOVERED intent is still there. But as the physical NFC tag isn't there anymore this leads to an error which is shown to the user. Setting the intent null in the Activity is not the best solution as there are situations where the Activity is re-created and with it the intent...
See also https://stackoverflow.com/questions/4116110/clearing-intent